Create hankerrank: max_colour_per_budget
Quuestion: A company is repainting it's office and would like to choose colours that work well together. They are presented with several various priced colours options presented in a specific order sod that each colour complement the surrounding colours. The company has a limited budget and would like to select the greatest number of consecutive colours that fit within the budget. Given the prices of the colours and the company budget, what is the maximum number of colours that can be purchased for this repainting project?